rename classes for quantinuum
